cpp05 compiles

+++ cpp06/ex01/include/Serializer.hpp
@@ -17,6 +17,6 @@ public:
 	~Serializer();
 	Serializer &operator=(const Serializer &src);
 
-	static uintptr_t serialize(Data* ptr);
-	static Data* deserialize(uintptr_t raw);
+	static unsigned long serialize(Data* ptr);
+	static Data* deserialize(unsigned long raw);
 };
