+++ cpp06/ex01/src/Serializer.cpp
@@ -15,11 +15,11 @@ Serializer::~Serializer()
 	std::cout << "Serializer Deconstructor called" << std::endl;
 }
 
-uintptr_t Serializer::serialize(Data* ptr)
+unsigned long Serializer::serialize(Data* ptr)
 {
-	return reinterpret_cast<uintptr_t>(ptr);
+	return reinterpret_cast<unsigned long>(ptr);
 }
-Data* Serializer::deserialize(uintptr_t raw)
+Data* Serializer::deserialize(unsigned long raw)
 {
 	return reinterpret_cast<Data*>(raw);
 }
